requirements.txt
pytesseract==0.3.10
PyQt5==5.15.9
Pillow==12.1.1
pdf2image==1.16.3
python-docx==1.1.0
pyinstaller==6.3.0
easyocr>=1.7.0
paddlepaddle>=2.6.0
paddleocr>=2.7.0
